Pulling IngestionJob out into its own module
